Move some OUtil stuff into separate header
